requirement.txt
selenium
python-dotenv

